# Conflicts:
#	crc/services/study_service.py

+++ crc/services/workflow_service.py
@@ -95,31 +95,38 @@ class WorkflowService(object):
             WorkflowService.delete_test_data()
             raise ApiError.from_workflow_exception("workflow_validation_exception", str(we), we)
 
+        count = 0
         while not processor.bpmn_workflow.is_completed():
-            try:
-                processor.bpmn_workflow.get_deep_nav_list()  # Assure no errors with navigation.
-                processor.bpmn_workflow.do_engine_steps()
-                tasks = processor.bpmn_workflow.get_tasks(SpiffTask.READY)
-                for task in tasks:
-                    if task.task_spec.lane is not None and task.task_spec.lane not in task.data:
-                        raise ApiError.from_task("invalid_role",
-                                       f"This task is in a lane called '{task.task_spec.lane}', The "
-                                       f" current task data must have information mapping this role to "
-                                       f" a unique user id.", task)
-                    task_api = WorkflowService.spiff_task_to_api_task(
-                        task,
-                        add_docs_and_forms=True)  # Assure we try to process the documentation, and raise those errors.
-                    # make sure forms have a form key
-                    if hasattr(task_api, 'form') and task_api.form is not None and task_api.form.key == '':
-                        raise ApiError(code='missing_form_key',
-                                       message='Forms must include a Form Key.',
-                                       task_id=task.id,
-                                       task_name=task.get_name())
-                    WorkflowService.populate_form_with_random_data(task, task_api, required_only)
-                    processor.complete_task(task)
-            except WorkflowException as we:
-                WorkflowService.delete_test_data()
-                raise ApiError.from_workflow_exception("workflow_validation_exception", str(we), we)
+            if count < 100:  # check for infinite loop
+                try:
+                    processor.bpmn_workflow.get_deep_nav_list()  # Assure no errors with navigation.
+                    processor.bpmn_workflow.do_engine_steps()
+                    tasks = processor.bpmn_workflow.get_tasks(SpiffTask.READY)
+                    for task in tasks:
+                        if task.task_spec.lane is not None and task.task_spec.lane not in task.data:
+                            raise ApiError.from_task("invalid_role",
+                                           f"This task is in a lane called '{task.task_spec.lane}', The "
+                                           f" current task data must have information mapping this role to "
+                                           f" a unique user id.", task)
+                        task_api = WorkflowService.spiff_task_to_api_task(
+                            task,
+                            add_docs_and_forms=True)  # Assure we try to process the documentation, and raise those errors.
+                        # make sure forms have a form key
+                        if hasattr(task_api, 'form') and task_api.form is not None and task_api.form.key == '':
+                            raise ApiError(code='missing_form_key',
+                                           message='Forms must include a Form Key.',
+                                           task_id=task.id,
+                                           task_name=task.get_name())
+                        WorkflowService.populate_form_with_random_data(task, task_api, required_only)
+                        processor.complete_task(task)
+                    count += 1
+                except WorkflowException as we:
+                    WorkflowService.delete_test_data()
+                    raise ApiError.from_workflow_exception("workflow_validation_exception", str(we), we)
+            else:
+                raise ApiError.from_task(code='validation_loop',
+                                         message=f'There appears to be an infinite loop in the validation. Task is {task.task_spec.description}',
+                                         task=task)
 
         WorkflowService.delete_test_data()
         WorkflowService._process_documentation(processor.bpmn_workflow.last_task.parent.parent)
